+++ templates/observables/l_to_l_conversion.cpp.in
@@ -30,7 +30,7 @@
 #include "cxx_qft/@ModelName@_qft.hpp"
 
 #include "@ModelName@_@filename@.hpp"
-#include "observables/ltolconversion/settings.hpp"
+#include "observables/l_to_l_conversion/settings.hpp"
 #include "@ModelName@_FFV_form_factors.hpp"
 
 #define F4F(quark, g) embed_photon<quark, Photon>(model, form_factors, g)
