+++ templates/slha_io.hpp.in
@@ -26,7 +26,7 @@
 #include "decays/flexibledecay_problems.hpp"
 #include "slha_io.hpp"
 #include "for_each.hpp"
-#include "observables/ltolconversion/settings.hpp"
+#include "observables/l_to_l_conversion/settings.hpp"
 #include "decays/decay.hpp"
 #include "decays/flexibledecay_settings.hpp"
